+++ google-cloud/fedora-coreos/kubernetes/workers/variables.tf
@@ -59,7 +59,7 @@ variable "preemptible" {
 
 variable "snippets" {
   type        = list(string)
-  description = "Fedora CoreOS Config snippets"
+  description = "Butane snippets"
   default     = []
 }